## Thoughts on using git for thoughts
[Steve Krouse](https://github.com/stevekrouse/futureofcoding.org) has an
amazing way of writing down all his thoughts in git commits. Makes me
think I could try out doing that for a while, as a replacement for
RedNotebook. It looks more portable and exportable, though I'm not sure
everyday usability will be better.

He also makes pretty fine-grained plans for his days, which feels like
an effective way of keeping track of what you spend time on, and a way
to avoid too many distractions. Reminds me of how I tried to plan my
days during the end of the PhD, then stopped doing it when we were in
the chaos of starting life here in Lyon.

Writing feels good, I'm thinking I'll try to restart this habit.

I feel like most of my time is spent reading, and trying to manage the
never-stopping branches of thought and more things to read that appear
all the time. Seems like making more detailed plans for the day could
help with this.