## More thoughts after reading commits and thoughts of Steve Krouse
I really like the design of his website, especially the super simple
colour palette. Compared to mine, which is a bit too grayscale, it feels
quite ventilated.

Second, and this is similar to what I had felt when writing a lot in
RedNotebook, I'm probably not going to come back very often on these
notes. If it helps me, it's to get ideas out of me, and help me with
playing around with them, *at the moment of writing*. Writing creates a
slower process of thought which lets me come back on things I thought a
few seconds to a few minutes ago, and lets me make a better map of the
possibilities that I was exploring.