## Wrapping up deep dive on Dynamic Land
For the past 2 weeks I've been exploring all aspects possible of Dynamic
Land and Bret Victor's work. It's been fascinating, and having shared
that content with XMCA makes me hope the conversation and reflection
will continue. I'm seriously considering starting a place like this
here:

* I have a large a3 page of notes about what Dynamic Land inspired for
me.
* I've registered to a number of meetups where I could find like-minded
people, found thanks to the scraping of the Twitter followers of
DynamicLand.
* I started a detailed pro-con list to inform a decision on what to do
next: either try to continue in research or start a part-time job in the
private sector to fund free-form research the other half of the time; or
try to combine both (i.e. don't dropout entirely, or, if continuing in
research, carve out time to start these projects).

It feels very good to have done this: I feel I have infused in those
ideas for two weeks, and I'm starting to feel that I've wrapped my head
around the subject.
Now it's time to step back, acknowledge what I haven't
read, and move on with:

* dayjob
* a concrete project around the enactive approach to language
* deciding (or combining) whether the next step is that project or
diving into creating a Dynamic Land here. This is all in my a3 paper on
what's next, and I have a bunch of people to talk with about this,
starting with Stéphane Grumbach and Marton.

I have a lot of thoughts about how Realtalk OS could be implemented,
after reading a bunch of pages of code from the twitter feed. The
important pictures of code are in `content/static/dynamicland`, along
with the zine, BV's 2014 research agenda, and a shot of BV's library
(supposedly), which could be good inspiration for what to read next once
I'm through Mindstorms and Ingold's "The Perception of the Environment".
The notes about how to implement are on a3 paper with me.

I also stumbled upon [Michael Nielsen](http://cognitivemedium.com/) who wrote the
[Neural Networks and Deep
Learning](http://neuralnetworksanddeeplearning.com/) book. There are a
lot of good ideas and posts there, especially the stuff about ChalkTalk
and reimplementing it with a tinge of neural networks to make it more
evolvable. The other good point is that he doesn't write too many posts,
so it's possible to go through his ideas in dense forms by clicking the
10-or-so links on his home page. I like that. His work seems to be in
the ballpark of Mindstorms, Ken Perlin (ChalkTalk), and Bret Victor. He
has [an interesting paper on Distill](https://distill.pub/2017/aia/)
which illustrates that.

## Original plan for 29/08/2018 afternoon

* Have a break
* Read through BV's research agenda and jot down notes and
criticisms. This should be very useful to kickstart merging
these ideas with the enactive approach to language and thought.
* Re-read my initial markdown notes for the Ezequiel draft
* Re-read the a3 paper on Dynamic Land
* Jot down anything else that occurrs to me at that moment
* If time allows, start writing the Ezequiel draft from those notes

But I bumped into Stéphane Grumbach and seized the opportunity for a
chat with him, which lasted most of the afternoon, moving most of the
things to spill-over for another day. See below:

## Long conversation with [Stéphane Grumbach](https://who.rocq.inria.fr/Stephane.Grumbach/), [Clément Renaud](http://clementrenaud.com/), and their intern (whose name I forgot)

Talked about:

* The future of universities, creating a university or a school
* The publication system, Elsevier and Google Scholar moving to
paid recommendation once OA will be legally required

### Spill-over for another day

* Read through BV's research agenda and jot down notes and
criticisms. This should be very useful to kickstart merging
these ideas with the enactive approach to language and thought.
* Re-read my initial markdown notes for the Ezequiel draft
* Re-read the a3 paper on Dynamic Land
* Jot down anything else that occurrs to me at that moment
* If time allows, start writing the Ezequiel draft from those notes
* Finish re-reading the Cuffari et al. 2014 article on language
* Find out if I can use a working version of STEPS (the "universal
document sounds like something that might be useful) so as not to have
to read through the whole report. But don't spend more than 30 minutes
on this.