## Projecting oneself into the future
There is also no such ideal in research in France I feel: I don't
have the ideal of "becoming a researcher paid by CNRS". I could have
the ideal of finding this or that, or revolutionasing a field, but
"becoming a researcher in France" implies such a crummy path to get
to such doubtful conditions, with no garantee that I would be able to
revolutionise something on the way, that there is no way it attracts me
like an ideal (even though there are many positive sides).

He also mentioned a number of books or things to read or look at:

* [La gouvernance par les
nombres](https://www.college-de-france.fr/site/alain-supiot/La-gouvernance-par-les-nombres-film.htm), by Alain Supiot
(both a book and a film)
* [La part maudite](https://fr.wikipedia.org/wiki/La_Part_maudite), an essay by Georges Bataille
* A book on the Lawyers in the Weimar republic, and how they were used
to develop the new law by having them work on the technical legal
problems, but not on the political and moral problems
* [The Schuman Declarationa](https://en.wikipedia.org/wiki/Schuman_Decl
aration) in 1950