## Notes from 3/09 morning
Going back to work majority-time on nw2vec, so I might be losing a bit
of the context of these past two weeks' thoughts. This is the difficulty
with multi-contexting things, as I have decided to do it, without having
hard commitments that force to switch back and forth when it's time. So
it's important that even if today's nw2vec plan bleeds into the
afternoon, I spend at least an hour on my other two tasks: writng to
Ezequiel and writing to Dynamic Land.

It's funny how headphones have become so popular among programming jobs
recently: everywhere I see people programming, they have big
high-quality headphones on. It occurs to me that this lets programmers
exercice their ear, i.e. another human skill than staring at a tiny
rectangle and moving their fingers.

## Dynamic Land thoughts of the day

Reduce the amount of backstage code as much as possible, i.e. have the
greatest percentage of actual running code *on pages in the room*. Maybe
run on top of STEPS? So use a few libraries as possible, if we can bake
things denser/terser.

## Log for 3/09

* Resumed nw2vec with some cleaning, rounding up issues, relaunching
the long-running minibatch exploration -- the detail is in the repo
* Started making a minimal plan for my presentation on Friday

## Plan for 4/09

* Look at results from the night's BlogCatalog run
* Implement the simpler Gaussian decoder without `u`, and launch a new
BlogCatalog run
* Spend at most 30 minutes resuming the GCN reproductions
* Concretely start my Friday presentation, making slides
* Have lunch
* Continue with the slides
* At 3pm, spend one hour on the Ezequiel draft
* At 4pm, spend the remaining time on Dynamic Land