## Log for 4/09
Gosh, I spent the whole day debugging this BlogCatalog problem, and got
to the bottom of it. Unfortunately that means I didn't get my other
items done, so it's bleeding into tomorrow. Still, this involved:

* Implement the simpler Gaussian decoder without `u`, and launch a new
BlogCatalog run

I also made some progress on Infuse networking, as the CEO of a
company in Paris might be interested by the project, and we could be
talking about it more in a few days. Yay!

## Plan for 5/09

* Fix my previous commit which has shifted dates
* Look at results from the night's BlogCatalog run
* Do some cleaning up of the nw2vec runs
* Spend at most 30 minutes resuming the GCN reproductions
* Concretely start my Friday presentation, making slides
* Have lunch
* Continue with the slides
* At 3pm, spend one hour on the Ezequiel draft
* At 4pm, spend the remaining time on Dynamic Land