## Notes from 5/09
My plan for today had to be readjusted again, as we had a long meeting
this morning with Marton and Jacobo. The full detail of what I do on
nw2vec doesn't really need to go in this log, which should be more dense
and about the high-level ideas. So I'll go back to putting that on paper
for now.

Academia has a way of taking over your time without warning which I'm
really not a fan of. I didn't get around to write my Ezequiel draft, nor
spend some time on Dynamic Land, nor start my slides for Friday (though
today's work does go into that, it's not concrete slide-making).

Finished listening to https://futureofcoding.org/episodes/26 which was
really interesting! Especially Glen's comments about how, if his
previous projects succeeded big time, people would not necessarily be
any happier in the world; a thought which made him turn to
psychotherapy. It's great to hear that is it puts back all the projects
I've been reading about into perspective. Food for thought.

## Plan for 6/09

Spend the whole day on my slides for Friday, from the first minute until
done.