## Graph Embedding Day in La Doua
Quite a lot of interesting things happened here. I met several people
that I'm happy to know now, including Rémy Cazabet, Aleksandar
Bojchevski and Anton Tsitsulin. And I made it alive to the end of the
day!

### About feature-network embeddings

The talks and conversations with Aleksandar and Anton finished
convincing me that our current take at nw2vec might by ill-posed, in
that we're mixing representations that are already vectors that could be
used separately (node2vec + node features). In what cases this is useful
is not clear. It would also be more satisfying to work on principles
and theory, i.e. a proper way of formulating problems, and not only on
empirical implementations. Mikhail, for instance, always asks where are
graphs a fundamental structure (other than in social networks) that is
prior to any other measure: you don't really need graphs when you're
converting other data into a graph structure.

This is a recurrent theme in research I find: you start a project, and
it turns out it's not the silver bullet it seemed to be at the
beginning, and ambitions drastically drop all of a sudden. I think in
this case there are several possible outcomes:
* find out that you're wrong, i.e. it *is* a silver bullet
* reorient the question
* extract the value of the work that was already done in understanding
concepts that are being used in the literature

At any rate my exploratory work is still very useful in understanding
how GCN & Co behave in an auto-encoder. And there's no point in deciding
the issue of this project now, it should continue its exploration and
include this new insight. Also, the genrative aspect of VAE could come
in useful. This points to the importance of going back to the literature
and revise world views regularly. Finally, it just occurred to me that
in the case where there *is* some level of correlation between features
and network, a combined representation should be able to deduplicate
compared to separate representations (which might be what Marton was
getting at from the beginning).

### iCAP

Rémy also told me a bit more about the [Learning Lab
Lyon](https://twitter.com/LearningLabLyon), which looks like the
anti-model for Dynamic Land. In spite of that, the people who initiated
that might be interesting to talk to. It's mostly
[iCAP](https://icap.univ-lyon1.fr/) apparently, and I think it would be
good to write to them. I'm starting a draft to remind me of that for my
next Dynamic Land session.

### DSRT and others

Other conversations led me to [DSRT](https://dsrt.fr/), a company in
Lyon which is developping [peer.us](https://peer.us/), i.e. making a
living by creating bibliograpy services to researchers. The company
was hard to find, which is not a very good sign, and I'm always turned
down by the poor web implementation and design skills demonstrated by
French companies, but I think it's very much worth having coffee with
the people behind this.

Another discovery was [IX-labs](http://www.ix-labs.org/) and [Sylvain
Peyronnet](https://twitter.com/speyronnet), a private research lab
(whose topics I'm not interested in, but it's interesting that it
exists).

### General thoughts

I'm getting more and more psyched about an Enactive project around a
Dynamic Land-ish place. My mind was blown over the end of August, and it
still is: the Future of Coding #26 with Glench was a good way of gaining
perspective, and confirmed that any of this only makes sense in the
context of a stronger research project on Dynamic Media and
communication. This feeling is paralleled by wanting to reduce my time
isolated in front of a computer while not relinquishing all interaction
with the computing world. I think I'll spend some real writing/email
time on this on Monday, it's really something I find exciting.