## Funding (independent) research
I really liked the [Independent
research](https://nadiaeghbal.com/independent-research) post by Nadia
Eghbal. It connected with a lof of options I've been reviewing lately
in order to make research work for me. The fact that several (many?)
famous researchers in the 18th and 19th century were either self-funded
(wealthy) or funded by mécénat. Darwin for instance (something I had
alread come across, but had forgotten). She talks about the challenge
of external validation (which is harder if you don't have a famous
institution backing you), but the real problem is funding I find.

Another part that impacted me is YCR's emphasis on output, and not
publications or conference participation, as a metric. Seems like it's
very necessary, and I very much resonate with that.

Then http://worrydream.com/2017-12-30-alan/ popped up in my exploration
of the Future of Coding slack (see below), and wham it resonated again.
Bret's and Alan's thought on how to fund and what worked in the past
are very insightful, reinforcing the impression that academia as it
exists today is not a good place to make progress in questions that can
interest me. (I especially like the emphasis on "problem finding", not
"problem solving"). Last night I was thinking about how research could
be funded through a variety of channels (for me as a person), and that
I don't need to wed myself to any particular one: consulting "gigs",
regular grants, donations, savings after a high-pay job, whatever. But
reading about the conditions that existed at PARC, ARPA, and the like,
it does feel like what I have been thinking about up to now is rather
meagre: great breakthroughs take a team, and by that I mean not only
internet-or-publication-mediated collaboration, but a team that meets
regularly in person and works together in a shared space ("much of
the fundamental creative work in computing requires 8 or more people
working together"). The options I've been thinking about don't really
get anywhere close to what's needed to fund a team.

This is making me think of expanding the "freelance funding" model:
imagine a group of people who have the skills to be hired at high-pay
jobs (maybe not every person in the group, but enough to sutain what
comes next), and decide to take turns at funding each other. A structure
where trust is strong enough that you feel comfortable working for
n days at a job 1) that is not necessarily your dream job, and 2)
for which you share the pay in order to fund the part of the team
that's not working. Then you get to do research for a month without
any constraints. It sounds like it would be a lot of unwanted work to
sustain a group, and it requires a huge level of trust (that others
won't leave after having benefited, that you feel okay working while
others are doing what they want, etc.). But it could be worth running
some numbers, and compare to the time spent raising funds in academia.
The plus side is that it would be people-centred, not project-centred.
Kind of a research coop, or a research solidarity commitment. Might be
stupid, but I'd like to see the details.

Seeing the amounts that Dynamic Land is looking for to fund its initial
years is a bit daunting, but could be a good benchmark to evaluate
these funding ideas (it includes research staff, rent, and equipment
purchase).

## Discovered new communities around Dynamic Land ideas

I decided to have a look at the [Future of Coding
slack](https://futureprogramming.slack.com/), and something clicked.
The conversations are pretty interesting, especially those that mention
Dynamic Land or point to other communities. I was super happy to
stumble upon [tbd.cool](http://discuss.tbd.cool/), started by another
BV fan who dreamt about (and is still dreaming about it?) working
at HARC. There is a great thread on DL which I want to get into,
especially since it mentions the publicity and people-interactions
aspects of DL (versus the computing capabilities). Another community is
[SomethingNew](https://github.com/d-cook/SomethingNew/). I particularly
like this: "We thus discourage trying to agree on "one way" of doing
things, and want to try to recognize and foster all possible views,
values, approaches, etc.". Not one project, but a forum. I like the
[Resources](https://github.com/d-cook/SomethingNew/blob/master/Resources
.md) and
[Projects](https://github.com/d-cook/SomethingNew/blob/master/Projects.m
d) pages, and there is much more to read in that repo. Started watching
it, so I'll get updates from there.

This leaves me with some reading, and a number of people to write to in
order to move on:

* Laurent Guerby, about his story for starting and funding Tetaneutral.net
* Judicael, to see if he's interested in all this
* tbd DL thread, touching both on the publicity of DL, and on the people
interaction aspect
* future of coding, simply to introduce myself
* xmca list, to answer the replies I got about DL
* Alix Ducros
* iCAP