## Working vs. looking for funding
I once read that we should "fall in love with a problem, not a
solution". Dynamic Land in a way is a solution, to the problem exposed
in The Humane Representation of Thought. That's no secret! It just
reminds me that I'd be better off spending time on my problem, which
closely relates to Bret's problem, than spending time on trying to find
funds for a Dynamic Land here. Bret has a knack for identifying the
common thread that is woven through all the things here's interested in.
He creates implementations that fascinate him, reacting to something
that obviously annoys him, and after a few of those he connects them and
finds the common theme.

My problem revolves around "understanding each other", in all the subtle
ways that may or may not yet be possible. I'm excited about it because
I'm starting to glimpse how understanding works (thanks to the Enactive
approach), and how it could be transformed, created in other ways, using
the right interaction substrates and facilitators (of which Dynamic Land
is one very good instance). It excites me also because it's tied with
humane technology, which empowers and puts people control. This theme
cuts through a number of solutions which I find fascinating: Dynamic
Land, the Enactive approach to language and interaction, my Infuse
project, locally produced electronics (possibly less powerful), openly
understandable tech, control by the person.

I've been putting a lot of effort into finding different models for
funding a Dynamic Land of sorts, or independent research, or both and a
team with that, but ultimately I'll get my problem solved by working on
it, mulling over it a lot more, talking to people about it and finding
other interested people, and having something to showcase when the
opportunity arises (be it a foundation, people, a grant, whatever).
This is the way to make progress both in actions and in my posing the
problem. JM Mercier, my physics professor in prépa, always said that
laying down the problem is by far the most important step in solving
it. (This connects with Bret's notion of "problem finding", instead of
"problem solving".) Right now I've been taking the academic approach of
trying to find funding *before* dedicating myself to my problem. Things
will flow a lot better by doing it the other way around, since I
currently have the opportunity to do so. Once I have working things, and
people I'm discussing this with, I can dream a lot more concretely about
what I'd do if I had €5M over two years to fund a place and a team, and
about where to find that money.

This brings me to my current plan for progress:

* make progress on my Infuse implementation
* talk to many many more people about my problem, and Dynamic Land and
Infuse (starting with yesterday's list of communities)
* buy hardware to start creating Dynamic-landy things at home