## Alan Kay: The Future Doesn't Have to be Incremental
People are, on dimension 1:
95% instrumentalists (will only be interested in something if it serves
their current goals, and are very conservative about changing their
goals)
5% focused on ideas and tools (will change their goals if they see
something that's just interesting)

And on dimension 2:
85% extroverts, who need people to agree with them (i.e. society's opinion
is the most important)
15% inner, self opinion most important

To make money, get anthropologist Robert Brown's book on human
universals, pick one and make a technqlqgical amplifier for one
of them, that will make a product that's popular with people. Those are
the things that hardly need to be marketed.
https://www.youtube.com/watch?v=gTAghAJcO1o at 22m35s.

A way of imagining the future we would like is to stop looking at the
part of the past that brought us to our current present (i.e. the past
who's heritage we live in), and look at overlooked things in the past,
and see what future they could yield if they were our current base.
That's a way to do Problem Finding. This used to be funded in the US,
but is not anymore.

Another way is to imagine a thing that would be cool or important to
have 30 years from now. Then imagine what that thing would look like 15
years from now, i.e. when it would be in its very infancy. Then try to
get the power of what that thing is running on, *today*. For computing,
that means having a super computer and using it as if it was a laptop.

The way to get good UIs is to do experiments with hundreds of users, and
that's what they did at Xerox PARC.